+++ src/server/game/Battlegrounds/Zones/BattlegroundAV.cpp
@@ -21,7 +21,7 @@
 BattlegroundAV::BattlegroundAV()
 {
     BgObjects.resize(BG_AV_OBJECT_MAX);
-    BgCreatures.resize(AV_CPLACE_MAX+AV_STATICCPLACE_MAX);
+    BgCreatures.resize(static_cast<int>(AV_CPLACE_MAX)+static_cast<int>(AV_STATICCPLACE_MAX));
 
     for (uint8 i = 0; i < 2; i++)
     {
@@ -293,7 +293,7 @@ Creature* BattlegroundAV::AddAVCreature(uint16 cinfoid, uint16 type)
 {
     bool isStatic = false;
     Creature* creature = nullptr;
-    ASSERT(type <= AV_CPLACE_MAX + AV_STATICCPLACE_MAX);
+    ASSERT(type <= static_cast<int>(AV_CPLACE_MAX) + static_cast<int>(AV_STATICCPLACE_MAX));
     if (type >= AV_CPLACE_MAX) //static
     {
         type -= AV_CPLACE_MAX;
@@ -625,8 +625,8 @@ void BattlegroundAV::EventPlayerDestroyedPoint(BG_AV_Nodes node)
         RewardReputationToTeam(ownerId == TEAM_ALLIANCE ? 730 : 729, BG_AV_REP_TOWER, ownerId);
         RewardHonorToTeam(GetBonusHonorFromKill(BG_AV_KILL_TOWER), ownerId);
 
-        SpawnBGObject(BG_AV_OBJECT_TAURA_A_DUNBALDAR_SOUTH+ownerId+(2*tmp), RESPAWN_ONE_DAY);
-        SpawnBGObject(BG_AV_OBJECT_TFLAG_A_DUNBALDAR_SOUTH+ownerId+(2*tmp), RESPAWN_ONE_DAY);
+        SpawnBGObject(static_cast<int>(BG_AV_OBJECT_TAURA_A_DUNBALDAR_SOUTH)+static_cast<int>(ownerId)+(2*tmp), RESPAWN_ONE_DAY);
+        SpawnBGObject(static_cast<int>(BG_AV_OBJECT_TFLAG_A_DUNBALDAR_SOUTH)+static_cast<int>(ownerId)+(2*tmp), RESPAWN_ONE_DAY);
     }
     else
     {
@@ -635,7 +635,7 @@ void BattlegroundAV::EventPlayerDestroyedPoint(BG_AV_Nodes node)
         else
             SpawnBGObject(object+11, RESPAWN_IMMEDIATELY);
         SpawnBGObject(BG_AV_OBJECT_AURA_N_FIRSTAID_STATION+3*node, RESPAWN_ONE_DAY);
-        SpawnBGObject(BG_AV_OBJECT_AURA_A_FIRSTAID_STATION+ownerId+3*node, RESPAWN_IMMEDIATELY);
+        SpawnBGObject(static_cast<int>(BG_AV_OBJECT_AURA_A_FIRSTAID_STATION)+static_cast<int>(ownerId)+3*node, RESPAWN_IMMEDIATELY);
         PopulateNode(node);
         if (node == BG_AV_NODES_SNOWFALL_GRAVE) //snowfall eyecandy
         {
@@ -955,7 +955,7 @@ void BattlegroundAV::EventPlayerDefendsPoint(Player* player, uint32 object)
     if (!IsTower(node))
     {
         SpawnBGObject(BG_AV_OBJECT_AURA_N_FIRSTAID_STATION+3*node, RESPAWN_ONE_DAY);
-        SpawnBGObject(BG_AV_OBJECT_AURA_A_FIRSTAID_STATION+teamId+3*node, RESPAWN_IMMEDIATELY);
+        SpawnBGObject(static_cast<int>(BG_AV_OBJECT_AURA_A_FIRSTAID_STATION)+static_cast<int>(teamId)+3*node, RESPAWN_IMMEDIATELY);
     }
         // despawn old go
     SpawnBGObject(object, RESPAWN_ONE_DAY);
@@ -1071,7 +1071,7 @@ void BattlegroundAV::EventPlayerAssaultsPoint(Player* player, uint32 object)
         {
             //spawning/despawning of aura
             SpawnBGObject(BG_AV_OBJECT_AURA_N_FIRSTAID_STATION+3*node, RESPAWN_IMMEDIATELY); //neutral aura spawn
-            SpawnBGObject(BG_AV_OBJECT_AURA_A_FIRSTAID_STATION+prevOwnerId+3*node, RESPAWN_ONE_DAY); //teeamaura despawn
+            SpawnBGObject(static_cast<int>(BG_AV_OBJECT_AURA_A_FIRSTAID_STATION)+static_cast<int>(prevOwnerId)+3*node, RESPAWN_ONE_DAY); //teeamaura despawn
 
             RelocateDeadPlayers(BgCreatures[node]);
         }
@@ -1727,7 +1727,7 @@ void BattlegroundAV::ResetBGSubclass()
     InitNode(BG_AV_NODES_SNOWFALL_GRAVE, TEAM_NEUTRAL, false); //give snowfall neutral owner
 
     m_Mine_Timer=AV_MINE_TICK_TIMER;
-    for (uint16 i = 0; i < AV_CPLACE_MAX+AV_STATICCPLACE_MAX; i++)
+    for (uint16 i = 0; i < static_cast<int>(AV_CPLACE_MAX)+static_cast<int>(AV_STATICCPLACE_MAX); i++)
         if (BgCreatures[i])
             DelCreature(i);
 }
