+++ src/server/scripts/Spells/spell_dk.cpp
@@ -222,7 +222,7 @@ class spell_dk_raise_ally : public SpellScriptLoader
                         ghoul->SetModifierValue(UNIT_MOD_HEALTH, TOTAL_VALUE, (ghoul->GetStat(STAT_STAMINA) - ghoul->GetCreateStat(STAT_STAMINA))*10.0f);
 
                         // Power Energy
-                        ghoul->SetModifierValue(UnitMods(UNIT_MOD_POWER_START+POWER_ENERGY), BASE_VALUE, ghoul->GetCreatePowers(POWER_ENERGY));
+                        ghoul->SetModifierValue(UnitMods(static_cast<int>(UNIT_MOD_POWER_START)+static_cast<int>(POWER_ENERGY)), BASE_VALUE, ghoul->GetCreatePowers(POWER_ENERGY));
                         ghoul->UpdateAllStats();
                         ghoul->SetFullHealth();
